Revert "Merge pull request #498 from ywatanabe1989/feat/per-agent-telegrammer-state-dir"
This reverts commit 6dc7a3ce12ca969d7ba68eeed5e75c0e732c5e80, reversing
changes made to 8b4c18cd725f12381f115492bc4870e3c7cf3d34.
